refactor (messaging): move display related code into the LastMsg impl
